+++ packages/@tinacms/react-alerts/src/Alerts.tsx
@@ -73,6 +73,7 @@ const AlertContainer = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
+  z-index: var(--tina-z-index-3);
 `
 
 const AlertEntranceAnimation = keyframes`