+++ packages/@tinacms/react-core/src/use-subscribable.tsx
@@ -17,7 +17,10 @@ limitations under the License.
 */
 
 import * as React from 'react'
-import { Subscribable } from '@tinacms/core'
+
+interface Subscribable {
+  subscribe(callback: any): () => void
+}
 /**
  *
  * @param subscribable An object that can be subscribed to