+++ packages/@tinacms/react-toolbar/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.2](https://github.com/tinacms/tinacms/compare/@tinacms/react-toolbar@0.1.2-alpha.2...@tinacms/react-toolbar@0.1.2) (2020-05-11)
+
+**Note:** Version bump only for package @tinacms/react-toolbar
+
+
+
+
+
+## [0.1.2-alpha.2](https://github.com/tinacms/tinacms/compare/@tinacms/react-toolbar@0.1.2-alpha.1...@tinacms/react-toolbar@0.1.2-alpha.2) (2020-05-08)
+
+**Note:** Version bump only for package @tinacms/react-toolbar
+
+
+
+
+
+## [0.1.2-alpha.1](https://github.com/tinacms/tinacms/compare/@tinacms/react-toolbar@0.1.2-alpha.0...@tinacms/react-toolbar@0.1.2-alpha.1) (2020-05-08)
+
+**Note:** Version bump only for package @tinacms/react-toolbar
+
+
+
+
+
+## [0.1.2-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-toolbar@0.1.1...@tinacms/react-toolbar@0.1.2-alpha.0) (2020-05-08)
+
+**Note:** Version bump only for package @tinacms/react-toolbar
+
+
+
+
+
+## [0.1.1](https://github.com/tinacms/tinacms/compare/@tinacms/react-toolbar@0.1.1-alpha.0...@tinacms/react-toolbar@0.1.1) (2020-05-04)
+
+**Note:** Version bump only for package @tinacms/react-toolbar
+
+
+
+
+
+## [0.1.1-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-toolbar@0.1.0...@tinacms/react-toolbar@0.1.1-alpha.0) (2020-04-28)
+
+**Note:** Version bump only for package @tinacms/react-toolbar
+
+
+
+
+
+# [0.1.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-toolbar@0.1.0-alpha.5...@tinacms/react-toolbar@0.1.0) (2020-04-27)
+
+**Note:** Version bump only for package @tinacms/react-toolbar
+
+
+
+
+
+# [0.1.0-alpha.5](https://github.com/tinacms/tinacms/compare/@tinacms/react-toolbar@0.1.0-alpha.4...@tinacms/react-toolbar@0.1.0-alpha.5) (2020-04-24)
+
+
+### Features
+
+* add form actions to toolbar ([8e1fc34](https://github.com/tinacms/tinacms/commit/8e1fc34))
+
+
+
+
+
 # [0.1.0-alpha.4](https://github.com/tinacms/tinacms/compare/@tinacms/react-toolbar@0.1.0-alpha.3...@tinacms/react-toolbar@0.1.0-alpha.4) (2020-04-20)
 
 **Note:** Version bump only for package @tinacms/react-toolbar