+++ packages/next-tinacms-json/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.15](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.14...next-tinacms-json@0.3.15) (2020-05-12)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+## [0.3.14](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.14-alpha.3...next-tinacms-json@0.3.14) (2020-05-11)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+## [0.3.14-alpha.3](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.14-alpha.2...next-tinacms-json@0.3.14-alpha.3) (2020-05-08)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+## [0.3.14-alpha.2](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.14-alpha.1...next-tinacms-json@0.3.14-alpha.2) (2020-05-08)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+## [0.3.14-alpha.1](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.14-alpha.0...next-tinacms-json@0.3.14-alpha.1) (2020-05-08)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+## [0.3.14-alpha.0](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.13...next-tinacms-json@0.3.14-alpha.0) (2020-05-06)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+## [0.3.13](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.13-alpha.0...next-tinacms-json@0.3.13) (2020-05-04)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+## [0.3.13-alpha.0](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.12...next-tinacms-json@0.3.13-alpha.0) (2020-04-28)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+## [0.3.12](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.12-alpha.4...next-tinacms-json@0.3.12) (2020-04-27)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+## [0.3.12-alpha.4](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.12-alpha.3...next-tinacms-json@0.3.12-alpha.4) (2020-04-24)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
 ## [0.3.12-alpha.3](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.12-alpha.2...next-tinacms-json@0.3.12-alpha.3) (2020-04-20)